+++ packages/rpmtw_api_client/example/main.dart
@@ -1,9 +1,11 @@
+import 'dart:io';
+
 import 'package:rpmtw_api_client/rpmtw_api_client.dart';
 
 void main() async {
-  RPMTWApiClient.init(); // Initialize RPMTW API Client
+  RPMTWApiClient.init(development: true); // Initialize RPMTW API Client
   RPMTWApiClient apiClient = RPMTWApiClient.instance;
-  print(apiClient.baseUrl);
+  print(apiClient.apiBaseUrl);
   PasswordValidatedResult result =
       await apiClient.authResource.validPassword("12345"); // Valid password
   print(result.message);
@@ -13,4 +15,5 @@ void main() async {
 
   String ip = await apiClient.otherResource.getIP();
   print(ip);
+  exit(0);
 }