+++ packages/rpmtw_api_client/lib/rpmtw_api_client.dart
@@ -15,12 +15,14 @@ export 'src/models/minecraft/mod_side.dart';
 export 'src/models/minecraft/relation_mod.dart';
 export 'src/models/minecraft/rpmwiki/wiki_change_log.dart';
 export 'src/models/curseforge/curseforge_mod.dart';
+export 'src/models/cosmic_chat/cosmic_chat_message.dart';
 
 export 'src/resources/auth_resource.dart';
 export 'src/resources/storage_resource.dart';
 export 'src/resources/minecraft_resource.dart';
 export 'src/resources/other_resource.dart';
 export 'src/resources/curseforge_resource.dart';
+export 'src/resources/cosmic_chat_resource.dart';
 
 
 export 'src/utilities/exceptions.dart';