+++ packages/rpmtw_api_client/lib/src/resources/base_resource.dart
@@ -2,11 +2,11 @@ import 'package:http/http.dart';
 
 abstract class BaseResource {
   Client httpClient;
-  String baseUrl;
+  String apiBaseUrl;
   String? globalToken;
 
   BaseResource(
       {required this.httpClient,
-      required this.baseUrl,
+      required this.apiBaseUrl,
       required this.globalToken});
 }